refactor: rename local variables according to renamed types/selectors
