+++ visualization/app/codeCharta/ui/colorSettingsPanel/metricColorRangeSlider/metricColorRangeSlider.component.html
@@ -1,11 +1,11 @@
 <cc-range-slider
-	*ngIf="!(isDeltaState$ | async)"
-	[minValue]="(colorRange$ | async).min"
-	[maxValue]="(colorRange$ | async).max"
-	[currentLeftValue]="(colorRange$ | async).from"
-	[currentRightValue]="(colorRange$ | async).to"
-	[leftColor]="(sliderColors$ | async).leftColor"
-	[middleColor]="(sliderColors$ | async).middleColor"
-	[rightColor]="(sliderColors$ | async).rightColor"
-	[handleValueChange]="handleValueChange"
+  *ngIf="!(isDeltaState$ | async)"
+  [minValue]="(sliderValues$ | async).min"
+  [maxValue]="(sliderValues$ | async).max"
+  [currentLeftValue]="(sliderValues$ | async).from"
+  [currentRightValue]="(sliderValues$ | async).to"
+  [leftColor]="(sliderColors$ | async).leftColor"
+  [middleColor]="(sliderColors$ | async).middleColor"
+  [rightColor]="(sliderColors$ | async).rightColor"
+  [handleValueChange]="handleValueChange"
 ></cc-range-slider>
